+++ crates/client/eth/src/l1_gas_price.rs
@@ -2,6 +2,7 @@ use crate::client::EthereumClient;
 use alloy::eips::BlockNumberOrTag;
 use alloy::providers::Provider;
 use anyhow::Context;
+use bigdecimal::BigDecimal;
 use mc_mempool::{GasPriceProvider, L1DataProvider};
 use std::time::{Duration, UNIX_EPOCH};
 
@@ -66,6 +67,21 @@ async fn update_gas_price(eth_client: &EthereumClient, l1_gas_provider: GasPrice
 
     l1_gas_provider.update_eth_l1_gas_price(*eth_gas_price);
     l1_gas_provider.update_eth_l1_data_gas_price(avg_blob_base_fee);
+
+    // fetch eth/strk price and update
+    if let Some(oracle_provider) = &l1_gas_provider.oracle_provider {
+        let (eth_strk_price, decimals) =
+            oracle_provider.fetch_eth_strk_price().await.expect("failed to retrieve ETH/STRK price");
+        let strk_gas_price = (BigDecimal::new((*eth_gas_price).into(), decimals.into())
+            / BigDecimal::new(eth_strk_price.into(), decimals.into()))
+        .as_bigint_and_exponent();
+        let strk_data_gas_price = (BigDecimal::new(avg_blob_base_fee.into(), decimals.into())
+            / BigDecimal::new(eth_strk_price.into(), decimals.into()))
+        .as_bigint_and_exponent();
+
+        l1_gas_provider.update_strk_l1_gas_price(strk_gas_price.0.to_str_radix(10).parse::<u128>()?);
+        l1_gas_provider.update_strk_l1_data_gas_price(strk_data_gas_price.0.to_str_radix(10).parse::<u128>()?);
+    }
 
     l1_gas_provider.update_last_update_timestamp();